+++ src/audio/createAudioStream.js
@@ -1,58 +1,41 @@
-const get = require('lodash/get');
 const { throwError } = require('rxjs');
-const {map,mergeMap,bufferTime} = require('rxjs/operators');
-const {client} = require('@buccaneerai/graphql-sdk');
+const {map,bufferTime} = require('rxjs/operators');
 const {toLinear16} = require('@buccaneerai/rxjs-linear16');
+const logger = require('@buccaneerai/logging-utils');
 
-const streamS3Audio = require('./streamS3Audio');
+const ingestAudioFromStorage = require('./ingestAudioFromStorage');
 const ingestAudioFromClient = require('./ingestAudioFromClient');
 const ingestFromPriorRun = require('./ingestFromPriorRun');
 // const audioChunkToLinear16 = require('./audioChunkToLinear16');
 
 const errors = {
-  audioFileDNE: () => new Error('audio file does not exist'),
   unknownInputType: () => new Error('unknown input type'),
 };
 
-const gql = (url = process.env.GRAPHQL_URL, token = process.env.JWT_TOKEN) => (
-  client({url, token})
-);
-
-const createInputStream = function createInputStream(
+const createInputStream = function createInputStream({
   config,
-  _findAudioFiles = gql().findAudioFiles,
+  _ingestAudioFromStorage = ingestAudioFromStorage,
   _ingestAudioFromClient = ingestAudioFromClient,
   _ingestFromPriorRun = ingestFromPriorRun,
-  _toLinear16 = toLinear16
-) {
+  _toLinear16 = toLinear16,
+  _logger = logger
+}) {
   return clientStream$ => {
     switch (config.inputType) {
+      // Stream data from an existing audio file
       case 's3File':
-        return _findAudioFiles({filter: {_id: config.audioFileId}}).pipe(
-          map(response => get(response, 'audioFiles[0]', null)),
-          mergeMap(audioFile =>
-            audioFile
-            ? streamS3Audio({
-              s3Bucket: audioFile.s3Bucket,
-              s3Key: audioFile.s3Key,
-              // s3Key: `${s3Prefix}${config.audioFileId}.linear16`,
-            })
-            : throwError(errors.audioFileDNE())
-          )
+        _logger.info(
+          'createInputStream.s3File',
+          {runId: config.runId, audioFileId: config.audioFileId}
         );
+        return _ingestAudioFromStorage({config});
       // Stream of audio data. For example, from a web brower.
       case 'audioStream':
-        return clientStream$.pipe(
-          _ingestAudioFromClient(),
-          // TODO: currently only supports raw LINEAR16 (mimeType=audio/l16)
-          // _audioChunkToLinear16({
-          //   audioEncoding: config.audioEncoding,
-          //   sampleRate: config.sampleRate,
-          //   channels: config.channels,
-          // })
-        );
+        _logger.info('createInputStream.audioStream', {runId: config.runId});
+        return clientStream$.pipe(_ingestAudioFromClient());
       // From our internal telephone API
       case 'telephoneCall':
+        _logger.info('createInputStream.telephoneCall', {runId: config.runId});
         return clientStream$.pipe(
           _ingestAudioFromClient(),
           bufferTime(1000),
@@ -64,7 +47,9 @@ const createInputStream = function createInputStream(
             firstChunkContainsHeaders: false,
           }),
         );
+      // Re-process an existing run
       case 'rerun':
+        _logger.info('createInputStream.rerun', {runId: config.runId});
         return clientStream$.pipe(
           _ingestFromPriorRun({runId: config.runId})
         );