+++ CHANGELOG.md
@@ -1,5 +1,46 @@
-# [1.6.0](https://github.com/buccaneerai/notestream-api/compare/v1.5.1...v1.6.0) (2023-01-26)
+# [1.7.0-dev.1](https://github.com/buccaneerai/notestream-api/compare/v1.6.0...v1.7.0-dev.1) (2023-04-04)
 
+
+### Bug Fixes
+
+* issues with token passing for runJobs ([51de487](https://github.com/buccaneerai/notestream-api/commit/51de4873d6871cceedd7e30df029391bfced9226))
+* resolved issue with STT pipeline where it did not use s3File input ([bd6123b](https://github.com/buccaneerai/notestream-api/commit/bd6123be25202b14f69cb721351e28299b77550c))
+* run folder ([7b84e41](https://github.com/buccaneerai/notestream-api/commit/7b84e41d2cf146fc4c89bfc49d9178acbfb350fc))
+* typo ([ed60dee](https://github.com/buccaneerai/notestream-api/commit/ed60dee5f27951765abb33ef5e7c051aea26478b))
+
+
+### Features
+
+* Add fathom engine ([9e63aac](https://github.com/buccaneerai/notestream-api/commit/9e63aac4ef22563f9d7b1c7f0aebe456cc666dfb))
+* added new HTTP api to create batch run job ([bb9a022](https://github.com/buccaneerai/notestream-api/commit/bb9a0222ea7ff23aeb6da191298a53df0ee78543))
+
+# [1.6.0-dev.6](https://github.com/buccaneerai/notestream-api/compare/v1.6.0-dev.5...v1.6.0-dev.6) (2023-02-27)
+
+### Bug Fixes
+
+* run folder ([7b84e41](https://github.com/buccaneerai/notestream-api/commit/7b84e41d2cf146fc4c89bfc49d9178acbfb350fc))
+
+# [1.6.0-dev.5](https://github.com/buccaneerai/notestream-api/compare/v1.6.0-dev.4...v1.6.0-dev.5) (2023-02-21)
+
+
+### Bug Fixes
+
+* issues with token passing for runJobs ([51de487](https://github.com/buccaneerai/notestream-api/commit/51de4873d6871cceedd7e30df029391bfced9226))
+* resolved issue with STT pipeline where it did not use s3File input ([bd6123b](https://github.com/buccaneerai/notestream-api/commit/bd6123be25202b14f69cb721351e28299b77550c))
+
+
+### Features
+
+* added new HTTP api to create batch run job ([bb9a022](https://github.com/buccaneerai/notestream-api/commit/bb9a0222ea7ff23aeb6da191298a53df0ee78543))
+
+# [1.6.0-dev.4](https://github.com/buccaneerai/notestream-api/compare/v1.6.0-dev.3...v1.6.0-dev.4) (2023-02-17)
+
+
+### Bug Fixes
+
+* typo ([ed60dee](https://github.com/buccaneerai/notestream-api/commit/ed60dee5f27951765abb33ef5e7c051aea26478b))
+
+# [1.6.0](https://github.com/buccaneerai/notestream-api/compare/v1.5.1...v1.6.0) (2023-01-26)
 
 ### Bug Fixes
 
@@ -12,6 +53,11 @@
 ### Features
 
 * added new rerun inputType for runs ([2f63561](https://github.com/buccaneerai/notestream-api/commit/2f63561af2c3d1fef4250bde9735251fa09e2bd6))
+
+
+### Features
+
+* Add fathom engine ([9e63aac](https://github.com/buccaneerai/notestream-api/commit/9e63aac4ef22563f9d7b1c7f0aebe456cc666dfb))
 
 # [1.6.0-dev.3](https://github.com/buccaneerai/notestream-api/compare/v1.6.0-dev.2...v1.6.0-dev.3) (2023-01-26)